# Implement TrustedRandomAccess for slice::{Chunks, ChunksMut, Windows}