As suggested by @bluss in https://github.com/rust-lang/rust/issues/47115#issuecomment-354888334