@sdroege 2018-01-03
Thanks for the fast review!

This is going to fail because of https://github.com/rust-lang/rust/pull/47113
I'll update it in a bit