@sdroege 2018-01-03
Rebased on top of latest master