@sdroege 2018-01-03
Good point, thanks. This should probably use `checked_add` and if it overflows then just use the length.

Will fix in a few hours, thanks for spotting this