@sdroege 2018-01-04
I fixed that now, but also see the discussion at https://github.com/rust-lang/rust/pull/47126#issuecomment-355149381 and the following 3 comments. It might make sense to reconsider the whole `TrustedRandomAccess` iterator or at least the zip implementation based on it.