@kennytm 2018-01-04
Thanks again! Just wanna confirm, will the expression `i * self.chunk_size` overflow?