@sdroege 2018-01-04
> Thanks again! Just wanna confirm, will the expression i * self.chunk_size overflow?

Yes, but only if you call it with an `i` that is not in range for the slice. AFAIU this is fine, that's what the "unchecked" / unsafe part of the trait is about: you need to ensure beforehand that the item you want to access does actually exist.